refactor: moved sbatch class with its parent
